SUSTAINING-690: Refactoring codebase
