+++ .gitignore
@@ -1,3 +1,5 @@
 .env
 .ruff_cache/
 __pycache__/
+.venv
+.DS_Store
