+++ slack_handlers/Dockerfile
@@ -1,9 +1,8 @@
-
 FROM python:3.9-slim
 
 WORKDIR /app
 
-COPY . /app
+COPY ./../ /app
 
 RUN pip install --no-cache-dir -r requirements.txt
